Inline `clocksource_845` into `clock_source`
I believe keeping the target-specific code small and inline is more
readable than having the sub-module, and this is the style I've already
been implementing in other code.
