+++ src/syscon/clock_source.rs
@@ -1,8 +1,5 @@
 //! Clock configuration for the peripherals
 
-#[cfg(feature = "845")]
-mod clocksource_845;
-
 use crate::syscon;
 
 /// Internal trait used configure peripheral clock sources
@@ -61,4 +58,47 @@ impl AdcClock {
     pub fn new_default() -> Self {
         Self { caldiv: 24, div: 0 }
     }
+}
+
+#[cfg(feature = "845")]
+mod target {
+    use crate::{
+        pac::syscon::fclksel::SEL_A,
+        syscon::{
+            self,
+            frg::{FRG, FRG0, FRG1},
+            IOSC,
+        },
+    };
+
+    use super::{PeripheralClock, PeripheralClockSelector};
+
+    macro_rules! peripheral_clocks {
+        (
+            $(
+                $clock:ty,
+                $sel:ident;
+            )*
+         ) => {
+            $(
+                impl PeripheralClock for $clock {
+                    const CLOCK: SEL_A = SEL_A::$sel;
+
+                    fn select<S>(_: &S, syscon: &mut syscon::Handle)
+                    where
+                        S: PeripheralClockSelector,
+                    {
+                        syscon.fclksel[S::REGISTER_NUM]
+                            .write(|w| w.sel().variant(Self::CLOCK));
+                    }
+                }
+            )*
+        };
+    }
+
+    peripheral_clocks!(
+        FRG<FRG0>, FRG0CLK;
+        FRG<FRG1>, FRG1CLK;
+        IOSC, FRO;
+    );
 }
